livekit-agent/requirements.txt
livekit-agents[google,openai]>=1.2.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
